+++ src/VGMInstrSet.h
@@ -25,7 +25,7 @@ public:
 		MENU_ITEM(VGMInstrSet, OnSaveAsSF2, L"Convert to SoundFont 2")
 	END_MENU()
 
-	VGMInstrSet(const std::string& format, RawFile* file, ULONG offset, ULONG length = 0, std::wstring name = L"VGMInstrSet",
+	VGMInstrSet(const std::string& format, RawFile* file, uint32_t offset, uint32_t length = 0, std::wstring name = L"VGMInstrSet",
 				VGMSampColl* theSampColl = NULL);
 	virtual ~VGMInstrSet(void);
 
@@ -36,7 +36,7 @@ public:
 	virtual bool GetInstrPointers();
 	virtual bool LoadInstrs();
 	
-	VGMInstr* AddInstr(ULONG offset, ULONG length, unsigned long bank, unsigned long instrNum, const std::wstring& instrName = L"");
+	VGMInstr* AddInstr(uint32_t offset, uint32_t length, unsigned long bank, unsigned long instrNum, const std::wstring& instrName = L"");
 
 	virtual FileType GetFileType() { return FILETYPE_INSTRSET; }
 
@@ -64,24 +64,24 @@ class VGMInstr :
 	public VGMContainerItem
 {
 public:
-	VGMInstr(VGMInstrSet* parInstrSet, ULONG offset, ULONG length, ULONG bank,
-			 ULONG instrNum, const std::wstring& name = L"Instrument");
+	VGMInstr(VGMInstrSet* parInstrSet, uint32_t offset, uint32_t length, uint32_t bank,
+			 uint32_t instrNum, const std::wstring& name = L"Instrument");
 	virtual ~VGMInstr(void);
 
 	virtual Icon GetIcon() { return ICON_INSTR; };
 
-	inline void SetBank(ULONG bankNum);
-	inline void SetInstrNum(ULONG theInstrNum);
+	inline void SetBank(uint32_t bankNum);
+	inline void SetInstrNum(uint32_t theInstrNum);
 	
 	VGMRgn* AddRgn(VGMRgn* rgn);
-	VGMRgn* AddRgn(ULONG offset, ULONG length, int sampNum, BYTE keyLow = 0,
-					BYTE keyHigh = 0x7F, BYTE velLow = 0, BYTE velHigh = 0x7F);
+	VGMRgn* AddRgn(uint32_t offset, uint32_t length, int sampNum, uint8_t keyLow = 0,
+					uint8_t keyHigh = 0x7F, uint8_t velLow = 0, uint8_t velHigh = 0x7F);
 
 	virtual bool LoadInstr();
 
 public:
-	ULONG bank;
-	ULONG instrNum;
+	uint32_t bank;
+	uint32_t instrNum;
  
 	VGMInstrSet* parInstrSet;
 	std::vector<VGMRgn*> aRgns;
